refactor: simplify setup
