--- api/package.json
@@ -1,9 +1,0 @@
-{
-  "private": true,
-  "name": "api",
-  "version": "0.0.0",
-  "dependencies": {
-    "graphql-request": "1.8.2",
-    "postmark": "2.3.4"
-  }
-}
