+++ api/send-confirmation-email.js
