+++ gatsby-node.js
@@ -1,3 +1,5 @@
+require('dotenv').config();
+
 exports.createPages = async ({ graphql, actions: { createPage } }) => {
   const {
     data: {
