+++ src/templates/EventTemplate.js
