--- www/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "www",
-  "private": true,
-  "version": "0.0.0",
-  "scripts": {
-    "build": "gatsby build",
-    "clean": "gatsby clean",
-    "develop": "gatsby develop",
-    "now-build": "yarn run build",
-    "now-dev": "yarn run develop -p $PORT"
-  },
-  "dependencies": {
-    "gatsby": "2.16.1",
-    "gatsby-plugin-postcss": "2.1.15",
-    "gatsby-plugin-purgecss": "4.0.1",
-    "gatsby-source-graphql": "2.1.20",
-    "graphcms-image": "1.1.0-beta4",
-    "react": "16.10.2",
-    "react-dom": "16.10.2"
-  },
-  "devDependencies": {
-    "postcss-preset-env": "6.7.0",
-    "tailwindcss": "1.1.3"
-  }
-}
